Remove version number from projectile dependency
It appears that Cask does not actually do anything with the version
number that one includes in the Caskfile. It is just there for the
sake of documentation. But, in my opinion, since the documentation
does not reflect reality, it is poor documentation at best and
misleading at worst. So I have decided to remove it.